news_aggregator/requirements.txt
feedparser==6.0.10
requests==2.28.2
pymongo==4.3.3
nltk==3.8.1
python-dotenv==1.0.0